feat: add ask ai endpoint

+++ models/knowledge.go
@@ -13,7 +13,7 @@ type Knowledge struct {
 	Embedding sql.NullString `gorm:"->:false;<-:create"`
 	CreatedAt time.Time      `gorm:"column:created_at;type:timestamp"`
 	UpdatedAt time.Time      `gorm:"column:updated_at;type:timestamp"`
-	Distance  int            `gorm:"->:create"`
+	Distance  float64        `gorm:"->:create"`
 }
 
 // TableName overrides the table name used by User to `profiles`
